% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differencing_and_dual_methods.R
\name{compute_rk}
\alias{compute_rk}
\title{Compute infinity norm to closed interval}
\usage{
compute_rk(y, v, D, k)
}
\arguments{
\item{v}{variable to be projected}

\item{lambda}{range variable}
}
\description{
Compute infinity norm to closed interval
}
