% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differencing_and_dual_methods.R
\name{dual_grad}
\alias{dual_grad}
\title{Compute dual objective gradient}
\usage{
dual_grad(y, v, D, lambda)
}
\arguments{
\item{y}{response}

\item{v}{dual}

\item{D}{differencing matrix}

\item{lambda}{range variable}
}
\description{
Compute dual objective gradient
}
