% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differencing_and_dual_methods.R
\name{duality_gap}
\alias{duality_gap}
\title{Compute duality gap}
\usage{
duality_gap(y, b, v, D, lambda)
}
\arguments{
\item{y}{response}

\item{b}{primal variable}

\item{v}{dual variable}

\item{D}{differencing matrix}

\item{lambda}{regularization parameter}
}
\description{
Compute duality gap
}
