% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_methods.R
\name{gcv}
\alias{gcv}
\title{Generalized Cross Validation}
\usage{
gcv(y, X, lambda)
}
\arguments{
\item{y}{response variables}

\item{X}{design matrix}

\item{lambda}{vector of tuning parameters}
}
\description{
\code{gcv} returns the leave-one-out
for a sequence of regularization parameter values. #'
}
