% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt_restrictions.R
\name{get_plot_restrictions_fcn}
\alias{get_plot_restrictions_fcn}
\title{Plot restrictions}
\usage{
get_plot_restrictions_fcn(fx, rx, nRow = 3, nCol = 3)
}
\arguments{
\item{fx}{handle to function in question}

\item{rx}{handle to function that returns a sequence of feasible sequence t and x+tv}

\item{nRow}{number of row plots}

\item{nCol}{number of column plots}
}
\description{
Plot restrictions
}
