% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_methods.R
\name{get_ridge_regression}
\alias{get_ridge_regression}
\title{Plot restrictions}
\usage{
get_ridge_regression(y, X, lambda)
}
\arguments{
\item{y}{response variable}

\item{X}{design matrix}

\item{lambda}{vector of tuning parameters}
}
\description{
\code{ridge_regression} returns the ridge regression coefficient estimates
for a sequence of regularization parameter values
}
