% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_desc.R
\name{gradient_descent_backtrack}
\alias{gradient_descent_backtrack}
\title{Gradient Descent (Backtracking Step-Size)}
\usage{
gradient_descent_backtrack(fx, gradf, x0, max_iter = 100, tol = 0.001)
}
\arguments{
\item{fx}{handle to function that returns objective function values}

\item{gradf}{handle to function that returns gradient of objective function #' @param x0 initial parameter estimate}

\item{max_iter}{maximum number of iterations}

\item{tol}{convergence tolerance}
}
\description{
Gradient Descent (Backtracking Step-Size)
}
