% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradient_desc.R
\name{gradient_step}
\alias{gradient_step}
\title{Gradient Step}
\usage{
gradient_step(gradf, t, x)
}
\arguments{
\item{gradf}{handle to function that returns gradient of objective function #' @param x current parameter estimate}

\item{t}{step-size}
}
\description{
Gradient Step
}
