% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differencing_and_dual_methods.R
\name{inf_norm}
\alias{inf_norm}
\title{Compute infinity norm to closed interval}
\usage{
inf_norm(v, lambda)
}
\arguments{
\item{v}{variable to be projected}

\item{lambda}{range variable}
}
\description{
Compute infinity norm to closed interval
}
