% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_methods.R
\name{k_fold_cv}
\alias{k_fold_cv}
\title{K-fold Cross Validation}
\usage{
k_fold_cv(y, X, lambda, k = 5, seed = 12345)
}
\arguments{
\item{y}{response variables}

\item{X}{design matrix}

\item{lambda}{vector of tuning parameters}

\item{k}{number of folds}

\item{seed}{seed for random number generator}
}
\description{
\code{k_fold_cv} returns K-fold cross-validation prediction error estimates #' for a sequence of regularization parameter values.
}
