% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differencing_and_dual_methods.R
\name{myGetDkn}
\alias{myGetDkn}
\title{Compute kth order differencing matrix}
\usage{
myGetDkn(k, n)
}
\arguments{
\item{k}{order of the differencing matrix}

\item{n}{Number of time points}
}
\description{
Compute kth order differencing matrix
}
