% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differencing_and_dual_methods.R
\name{nuclear_norm}
\alias{nuclear_norm}
\title{Compute infinity norm to closed interval}
\usage{
nuclear_norm(D)
}
\arguments{
\item{v}{variable to be projected}

\item{lambda}{range variable}
}
\description{
Compute infinity norm to closed interval
}
