% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differencing_and_dual_methods.R
\name{project}
\alias{project}
\title{Compute projection to a closed interval -k,k}
\usage{
project(z, lambda)
}
\arguments{
\item{z}{variable to be projected}

\item{lambda}{range variable}
}
\description{
Compute projection to a closed interval -k,k
}
