% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/differencing_and_dual_methods.R
\name{trend_filter_pg}
\alias{trend_filter_pg}
\title{Solve trend-filtering by proximal gradient on dual problem}
\usage{
trend_filter_pg(y, k, v, lambda = 10, max_iter = 100, tol = 0.001)
}
\arguments{
\item{y}{response}

\item{k}{order of differencing matrix}

\item{v}{Initial dual variable}

\item{lambda}{regularization parameter}

\item{max_iter}{maximum number of iterations}

\item{tol}{convergence tolerance}
}
\description{
Solve trend-filtering by proximal gradient on dual problem
}
